feat(54): refactor entity wrapping / path classification into a new common base class
