+++ tests/rest_api_v2_client_test.py
@@ -240,6 +240,7 @@ class FunctionDecoratorsTest(unittest.TestCase):
         do_http_things.__name__ = 'rget'
         dummy_session.url = 'https://api.pagerduty.com'
         dummy_session.canonical_path.return_value = '/users'
+        dummy_session.entity_wrappers.return_value = ('users', 'users')
         self.assertEqual(users_array,
             pagerduty.wrapped_entities(do_http_things)(dummy_session, '/users',
                 query='user'))
@@ -251,6 +252,7 @@ class FunctionDecoratorsTest(unittest.TestCase):
         user_payload = {'email':'user@example.com', 'name':'User McUserson'}
         dummy_session.url = 'https://api.pagerduty.com'
         dummy_session.canonical_path.return_value = '/users'
+        dummy_session.entity_wrappers.return_value = ('user', 'user')
         self.assertRaises(
             pagerduty.Error,
             pagerduty.wrapped_entities(do_http_things),
@@ -280,6 +282,7 @@ class FunctionDecoratorsTest(unittest.TestCase):
         incidents = [{'id':'PABC123'}, {'id':'PDEF456'}]
         dummy_session.url = 'https://api.pagerduty.com'
         dummy_session.canonical_path.return_value = '/incidents'
+        dummy_session.entity_wrappers.return_value = ('incidents', 'incidents')
         do_http_things.__name__ = 'rput'
         response.ok = True
         updated_incidents = copy.deepcopy(incidents)
